requirements.txt
django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
pillow>=10.0.0
mysqlclient>=2.1.0
django-rest-auth>=0.9.5
djangorestframework-simplejwt>=5.2.0
django-allauth>=0.50.0